feat: make command, controller and work first command

+++ src/main/java/Library/Utils/IDGenerator/IDGenerator.java
@@ -4,5 +4,5 @@ public interface IDGenerator {
 
   String nextId();
 
-  long nextNumberId();
+  Integer nextNumberId();
 }
